+++ cmd/evidra/multi_string_flag.go
@@ -1,0 +1,14 @@
+package main
+
+import "strings"
+
+type multiStringFlag []string
+
+func (f *multiStringFlag) String() string {
+	return strings.Join(*f, ", ")
+}
+
+func (f *multiStringFlag) Set(v string) error {
+	*f = append(*f, v)
+	return nil
+}